# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bitcoin AI Analyst Backend",
    "version": "2026-09-research-evidence-v2"
  },
  "paths": {
...
    "/api/Indexer/rebuild-derived": {
      "post": {
        "tags": [
          "Indexer"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Indexer/technical-indicators": {
      "post": {
        "tags": [
          "Indexer"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Indexer/ml-dataset": {
      "post": {
        "tags": [
          "Indexer"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}